C99 printf() -Werror fixes
git-svn-id: svn://svn.berlios.de/openocd/trunk@2336 b42882b7-edfa-0310-969c-e2dbd0fdcd60

+++ src/flash/lpc3180_nand_controller.c
@@ -750,11 +750,11 @@ static int lpc3180_read_page(struct nand_device_s *device, uint32_t page, uint8_
 			{
 				if (mlc_isr & 0x40)
 				{
-					LOG_ERROR("uncorrectable error detected: 0x%2.2x", mlc_isr);
+					LOG_ERROR("uncorrectable error detected: 0x%2.2x", (unsigned)mlc_isr);
 					return ERROR_NAND_OPERATION_FAILED;
 				}
 
-				LOG_WARNING("%i symbol error detected and corrected", ((mlc_isr & 0x30) >> 4) + 1);
+				LOG_WARNING("%i symbol error detected and corrected", ((int)(((mlc_isr & 0x30) >> 4) + 1)));
 			}
 
 			if (data)
